Fix : DB 구조 변경

+++ src/posts/post-comments/post-comments.service.ts
@@ -1,16 +1,18 @@
-import { BadRequestException, Injectable, NotFoundException, UnauthorizedException } from '@nestjs/common';
+import { Inject, Injectable, NotFoundException, UnauthorizedException } from '@nestjs/common';
 import { CreatePostCommentDto } from './dto/create-post-comment.dto';
 import { UpdatePostCommentDto } from './dto/update-post-comment.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { PostComments } from './entities/post-comments.entity';
 import { Repository } from 'typeorm';
-import { Posts } from '../entities/post.entity';
+import { Posts } from '../entities/posts.entity';
+import { Cache, CACHE_MANAGER } from '@nestjs/cache-manager';
 
 @Injectable()
 export class PostCommentsService {
   constructor( 
     @InjectRepository(Posts) private readonly postsRepository : Repository<Posts>,
-    @InjectRepository(PostComments) private postCommentsRepository : Repository<PostComments>
+    @InjectRepository(PostComments) private postCommentsRepository : Repository<PostComments>,
+    @Inject(CACHE_MANAGER) private cacheManager : Cache
   ){}
 
   // 해당 게시물 댓글 생성
@@ -27,74 +29,38 @@ export class PostCommentsService {
     const { context } = createPostCommentDto;
 
     const createPostComment = this.postCommentsRepository.create({
-      postId,
       userId,
+      postId,
       context
     });
-
+    
     await this.postCommentsRepository.save(createPostComment);
+
+    const cached = await this.cacheManager.get(`post:${postId}`);
+
+    if(!cached){
+      await this.cacheManager.del(`post:${postId}`);
+    }
     
     return { statusCode : 201, message : "성공적으로 노래 댓글 생성이 완료되었습니다." };
   }
 
-  // 해당 게시물 댓글 전체 조회
-  async findAll(postId : number, page : number, page_size : number) {
-    const findPostOne = await this.postsRepository.findOne({ 
-      where : { id : postId, isOpen : true }, 
-      select : ['id'] 
-    });
-
-    if(!findPostOne){
-      throw new NotFoundException("노래 목록이 존재하지 않습니다.");
-    }
-
-    if(!page){
-      page = 1;
-    }
-
-    if(!page_size){
-      page_size = 10;
-    }
-
-    const find = await this.postCommentsRepository.find({
-      where : { postId },
-      relations : { users : true },
-      select : {
-        id : true,
-        context : true,
-        createdAt : true,
-        users : {
-          id : true,
-          nickname : true,
-          image : true
-        },
-      }, 
-      skip : ((page - 1) * page_size),
-      take : page_size
-    });
-    
-    if(find && find.length === 0){
-      throw new NotFoundException("노래 댓글 목록들이 존재하지 않습니다.");
-    }
-    
-    return { statusCode : 200, message : "성공적으로 노래 댓글 전체 조회가 완료되었습니다.", data : find };
-  }
-
   // 해당 게시물 댓글 삭제 리스트 전체 조회 (어드민만 가능)
   async findDeletedList() {
-    const findComment = await this.postCommentsRepository.createQueryBuilder("post-comments")
+    const findComment = await this.postCommentsRepository.createQueryBuilder("post_comments")
     .withDeleted()
-    .where('post-comments.deletedAt IS NOT NULL')
-    .innerJoin('post-comments.users', 'users')
+    .where('post_comments.deletedAt IS NOT NULL')
+    .innerJoin('post_comments.users', 'users')
+    .innerJoin('users.userInfos', 'userInfos')
     .select([
-      'post-comments.id',
-      'post-comments.context',
-      'post-comments.createdAt',
-      'post-comments.updatedAt',
-      'post-comments.deletedAt',
+      'post_comments.id',
+      'post_comments.context',
+      'post_comments.createdAt',
+      'post_comments.updatedAt',
+      'post_comments.deletedAt',
       'users.id',
       'users.nickname',
-      'users.image'
+      'userInfos.image'
     ])
     .getMany();
 
@@ -103,39 +69,6 @@ export class PostCommentsService {
     }
 
     return { statusCode : 200, message : "성공적으로 삭제 예정된 노래 댓글 전체 조회가 완료되었습니다.", data : findComment };
-  }
-
-  // 해당 게시물 댓글 상세 조회
-  async findOne(postId: number, id : number) {
-    const findPostOne = await this.postsRepository.findOne({ 
-      where : { id : postId, isOpen : true }, 
-      select : ['id'] 
-    });
-
-    if(!findPostOne){
-      throw new NotFoundException("노래 목록이 존재하지 않습니다.");
-    }
-
-    const findOne = await this.postCommentsRepository.findOne({
-      where : { postId, id },
-      relations : { users : true },
-      select : {
-        id : true,
-        context : true,
-        createdAt : true,
-        users : {
-          id : true,
-          nickname : true,
-          image : true
-        },
-      },
-    });
-
-    if(!findOne){
-      throw new NotFoundException("노래 댓글 목록이 존재하지 않습니다.");
-    }
-
-    return { statusCode : 200, message : "성공적으로 노래 댓글 상세 조회가 완료되었습니다.", data : findOne };
   }
 
   // 해당 게시물 댓글 수정
@@ -149,7 +82,16 @@ export class PostCommentsService {
       throw new NotFoundException("노래 목록이 존재하지 않습니다.");
     }
 
-    if(findPostOne.userId !== userId){
+    const findOne = await this.postCommentsRepository.findOne({
+      where : { id, postId },
+      select : ['id', 'userId']
+    });
+
+    if(!findOne){
+      throw new NotFoundException("노래 댓글 목록이 존재하지 않습니다.");
+    }
+
+    if(findOne.userId !== userId){
       throw new UnauthorizedException("유저 정보가 일치하지 않아 수정이 불가능합니다.")
     }
 
@@ -158,6 +100,12 @@ export class PostCommentsService {
     await this.postCommentsRepository.update(id, {
       context
     });
+
+    const cached = await this.cacheManager.get(`post:${postId}`);
+
+    if(!cached){
+      await this.cacheManager.del(`post:${postId}`);
+    }
 
     return { statusCode : 201, message : "성공적으로 노래 댓글 수정이 완료되었습니다." };
   }
@@ -174,7 +122,22 @@ export class PostCommentsService {
       throw new NotFoundException("노래 목록이 존재하지 않습니다.");
     }
 
+    const findOne = await this.postCommentsRepository.findOne({
+      where : { id, postId },
+      select : ['id']
+    });
+
+    if(!findOne){
+      throw new NotFoundException("노래 댓글 목록이 존재하지 않습니다.");
+    }
+
     await this.postCommentsRepository.delete(id);
+
+    const cached = await this.cacheManager.get(`post:${postId}`);
+
+    if(!cached){
+      await this.cacheManager.del(`post:${postId}`);
+    }
 
     return { statusCode : 201, message : "성공적으로 노래 댓글 삭제가 완료되었습니다." };
   }
@@ -190,13 +153,28 @@ export class PostCommentsService {
       throw new NotFoundException("노래 목록이 존재하지 않습니다.");
     }
 
-    if(findPostOne.userId !== userId){
+    const findOne = await this.postCommentsRepository.findOne({
+      where : { id, postId },
+      select : ['id', 'userId']
+    });
+
+    if(!findOne){
+      throw new NotFoundException("노래 댓글 목록이 존재하지 않습니다.");
+    }
+
+    if(findOne.userId !== userId){
       throw new UnauthorizedException("유저 정보가 일치하지 않아 삭제가 불가능합니다.");
     }
 
     await this.postCommentsRepository.update(id, {
       deletedAt : new Date()
     });
+
+    const cached = await this.cacheManager.get(`post:${postId}`);
+
+    if(!cached){
+      await this.cacheManager.del(`post:${postId}`);
+    }
 
     return { statusCode : 201, message : "성공적으로 노래 댓글 삭제가 완료되었습니다." };
   }
